contracts/requirements.txt
beaker-pyteal>=0.2.0
algokit>=1.0.0
algokit-utils>=1.0.0
algosdk>=2.7.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
